- NEW FEATURE: Viewer option menu added for automatic scaling of all tracks and 3 different heights for track viewer - BUGFIX: Genetic codes only with starts or stops fixed

+++ ReadXplorer/ReadXplorer_Backend/src/de/cebitec/readXplorer/databackend/SamBamFileReader.java
@@ -343,7 +343,7 @@ public class SamBamFileReader implements Observable {
                                 startPos, stop, isFwdStrand, mappingQuality, record.getBaseQualities());
                         if (pairId != null && pairType != null) { //since both data fields are always written together
 //                // add new readPair if not exists
-                            readPairId = (long) pairId;
+                            readPairId = pairId;
                             readPairType = ReadPairType.getReadPairType(pairType);
                             if (!readPairs.containsKey(readPairId)) {
                                 newGroup = new PersistantReadPairGroup();
